DMT-40 Extract helper functions to a separate class
Refactor codebase to encapsulate the utility functions into a separate component.
